feat: 상품 CRUD 기본기능 구현 및 DTO로 반환

+++ src/main/java/com/back/project1_team1/product/dto/ProductResponse.java
@@ -3,10 +3,11 @@ package com.back.project1_team1.product.dto;
 import com.back.project1_team1.product.Product;
 
 public record ProductResponse(
-        int id,
+        Long id,
         String name,
         int price
 ) {
+    // Product Entity를 ProductResponse DTO로 변환하는 생성자
     public ProductResponse(Product product) {
         this(
                 product.getId(),
